+++ tests/21-tag-with-branch-config.sh
@@ -1,22 +1,17 @@
 #!/bin/bash -e
-# Test tag config parameters
+# Test branch with config
 h="$PWD"
 d=$1
 shift
 cd "$d"
-
-# Init GNUPG
 export GNUPGHOME="$d/gnupg"
 mkdir -p -m 700 "$GNUPGHOME"
+git init
 
-# Init GIT
-if [ ! -d .git ]; then
-	git init
-fi
-echo $RANDOM > a.txt
-git add a.txt
-git commit -m "Random change $RANDOM"
-tagid=v$RANDOM
+echo $RANDOM > 21-a.txt
+git add 21-a.txt
+git commit -m "Random change 21-$RANDOM"
+tagid=v21-$RANDOM
 
 # Change config
 git config timestamp.server https://gitta.zeitgitter.net