+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## Added
 
 ## Fixed
+- No longer allow timestamping a timestamp branch head onto itself
+  (would cause a merge commit with the same parent twice)
 
 ## Changed
 - Updated and extended [ServerList.md](./doc/ServerList.md)