+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
-# 0.9.7+ - [Unreleased]
+# 1.0.0 - 2019-09-23
 ## Added
 - Can now run selected tests only: `make tests TESTS=tests/20-branch.sh` etc.
 - `--append-branch-name` (default: `yes`) appends the branch name on